backend/requirements.txt
# OBIE Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# ORM Helpers
pydantic==2.5.3
pydantic-settings==2.1.0

# Task Queue
celery==5.3.6
redis==5.0.1
flower==2.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Web Scraping (use sparingly)
playwright==1.40.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.3
lxml==5.1.0

# RSS Feeds
feedparser==6.0.11

# Email Validation
dnspython==2.5.0

# LLM (Groq for speed)
groq==0.4.2

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3
faker==21.0.0

# Observability
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
respx==0.20.2

# Dev Tools
black==24.1.0
ruff==0.1.14
mypy==1.8.0
